+++ full-service/src/json_rpc/v1/api/wallet.rs
@@ -39,9 +39,10 @@ use crate::{
         confirmation_number::ConfirmationService,
         gift_code::{EncodedGiftCode, GiftCodeService},
         ledger::LedgerService,
+        models::transaction_memo::TransactionMemo,
         payment_request::PaymentRequestService,
         receipt::ReceiptService,
-        transaction::{TransactionMemo, TransactionService},
+        transaction::TransactionService,
         transaction_log::TransactionLogService,
         txo::TxoService,
         WalletService,
