+++ full-service/src/json_rpc/v2/api/wallet.rs
@@ -39,12 +39,12 @@ use crate::{
         hardware_wallet::sync_txos,
         ledger::LedgerService,
         memo::MemoService,
-        models::tx_proposal::TxProposal,
+        models::{transaction_memo::TransactionMemo, tx_proposal::TxProposal},
         network::get_token_metadata,
         payment_request::PaymentRequestService,
         receipt::ReceiptService,
         signed_contingent_input::SignedContingentInputService,
-        transaction::{TransactionMemo, TransactionService},
+        transaction::TransactionService,
         transaction_log::TransactionLogService,
         txo::TxoService,
         watcher::WatcherService,
