+++ full-service/src/service/gift_code.rs
@@ -20,8 +20,8 @@ use crate::{
         account::AccountServiceError,
         address::{AddressService, AddressServiceError},
         ledger::{LedgerService, LedgerServiceError},
-        models::tx_proposal::TxProposal,
-        transaction::{TransactionMemo, TransactionService, TransactionServiceError},
+        models::{transaction_memo::TransactionMemo, tx_proposal::TxProposal},
+        transaction::{TransactionService, TransactionServiceError},
         transaction_builder::DEFAULT_NEW_TX_BLOCK_ATTEMPTS,
         WalletService,
     },
