+++ full-service/src/service/models/mod.rs
@@ -1,4 +1,5 @@
 pub mod ledger;
+pub mod transaction_memo;
 pub mod tx_blueprint_proposal;
 pub mod tx_proposal;
 pub mod watcher;
