+++ full-service/src/service/transaction_log.rs
@@ -126,9 +126,8 @@ mod tests {
         db::account::AccountID,
         json_rpc::v2::models::amount::Amount,
         service::{
-            account::AccountService,
-            address::AddressService,
-            transaction::{TransactionMemo, TransactionService},
+            account::AccountService, address::AddressService,
+            models::transaction_memo::TransactionMemo, transaction::TransactionService,
             transaction_log::TransactionLogService,
         },
         test_utils::{
