+++ full-service/src/service/txo.rs
@@ -16,8 +16,8 @@ use crate::{
     json_rpc::v2::models::amount::Amount,
     service::{
         ledger::LedgerServiceError,
-        models::tx_proposal::TxProposal,
-        transaction::{TransactionMemo, TransactionService, TransactionServiceError},
+        models::{transaction_memo::TransactionMemo, tx_proposal::TxProposal},
+        transaction::{TransactionService, TransactionServiceError},
     },
     WalletService,
 };
